MSI: 379: Product with negative Quantity doesn't change Stock status to "Out of stock"
