+++ app/code/Magento/Inventory/composer.json
@@ -10,8 +10,6 @@
         "magento/module-backend": "100.3.*",
         "magento/module-directory": "100.3.*",
         "magento/module-shipping": "100.3.*",
-        "magento/module-catalog-inventory": "100.3.*",
-        "magento/module-catalog": "101.2.*",
         "magento/module-ui": "100.3.*"
     },
     "type": "magento2-module",
