+++ app/code/Magento/InventoryConfiguration/composer.json
@@ -6,6 +6,7 @@
         "magento/framework": "100.3.*",
         "magento/module-inventory": "100.0.*",
         "magento/module-inventory-api": "100.0.*",
+        "magento/module-catalog": "100.3.*",
         "magento/module-catalog-inventory": "100.3.*"
     },
     "type": "magento2-module",
